refactor: split alarm configuration routes into dedicated alarm_config module for improved organization
